move world-symbols to file_symbols
